+++ xtransport/lib/src/ws_html.dart
@@ -34,14 +34,14 @@ class XTransportWsClient implements ITransportClient {
   //Events
   void Function()? _onConnect;
   void Function()? _onClose;
-  void Function(Error err)? _onError;
+  void Function(XTransportError err)? _onError;
   void Function(Message msg)? _onMessage;
 
   //Method
   @override
   void send(ITransportPacket obj) {
     try {
-      _socket?.send(obj.pack());
+      _socket?.sendTypedData(obj.pack());
     } catch (e) {
       if (log) {
         loger.log(
@@ -70,7 +70,7 @@ class XTransportWsClient implements ITransportClient {
   void onConnect(void Function() fn) => _onConnect = fn;
 
   @override
-  void onError(void Function(Error err) fn) => _onError = fn;
+  void onError(void Function(XTransportError err) fn) => _onError = fn;
 
   @override
   void onMessage(void Function(Message msg) fn) => _onMessage = fn;
@@ -141,7 +141,7 @@ class XTransportWsClient implements ITransportClient {
     } catch (e) {
       if (log) loger.log("connect error: $e", name: "ws");
       status = ConnectStatus.disconnect;
-      _onError?.call(Error.from(e));
+      _onError?.call(XTransportError.from(e));
       _onClose?.call();
       return Future.value();
     }
@@ -214,7 +214,7 @@ class XTransportWsClient implements ITransportClient {
     _onErrorSubscription = _socket?.onError.listen((e) {
       if (log) loger.log("onError", name: "ws", error: e.toString());
       status = ConnectStatus.disconnect;
-      _onError?.call(Error.from(e));
+      _onError?.call(XTransportError.from(e));
     });
     _onOpenSubscription = _socket?.onOpen.listen((e) {
       status = ConnectStatus.connected;