+++ xtransport/lib/xtransport.dart
@@ -2,5 +2,6 @@ library xtransport;
 
 export 'src/shared/credentials.dart';
 export 'src/interface.dart';
+export 'src/jsons.dart';
 export 'src/tcp.dart';
 export 'src/ws.dart' show XTransportWsClient;